add cache to the serializer
